fix(user-group): also add permissions below each perm group on edit page
